修改类型转换的计算位置

+++ Assets/JobsTutorials/Lesson4/Scripts/WaveCubeEntityMoveJob.cs
@@ -6,12 +6,13 @@ using Unity.Transforms;
 namespace DOTS.DOD.Lesson4
 {
     [BurstCompile]
+    
     partial struct WaveCubeEntityMoveJob : IJobEntity
     {
-        public float elapsedTime;
+        public double elapsedTime;
         void Execute(ref LocalTransform localTransform,in WaveCubeDistance cubeDistance)
         {
-            float y = math.sin(elapsedTime * 3f + cubeDistance.distance * 0.2f) * 9;
+            float y = (float)math.sin(elapsedTime * 3f + cubeDistance.distance * 0.2f) * 9;
             localTransform.Position = new float3(localTransform.Position.x, y, localTransform.Position.z);
         }
     }
